requirements.txt
streamlit
biopython
pandas
matplotlib
requests